+++ x/ccv/consumer/types/keys_test.go
@@ -41,6 +41,7 @@ func getAllKeyPrefixes() []byte {
 		InitGenesisHeightByteKey,
 		StandaloneTransferChannelIDByteKey,
 		PrevStandaloneChainByteKey,
+		PendingPacketsIndexByteKey,
 		BouncingSlashByteKey,
 		RetryAllowedByteKey,
 	}
@@ -74,11 +75,12 @@ func getAllFullyDefinedKeys() [][]byte {
 		PacketMaturityTimeKey(0, time.Time{}),
 		HeightValsetUpdateIDKey(0),
 		OutstandingDowntimeKey([]byte{}),
-		PendingDataPacketsKey(),
+		PendingDataPacketsKey(473289),
 		CrossChainValidatorKey([]byte{}),
 		InitGenesisHeightKey(),
 		StandaloneTransferChannelIDKey(),
 		PrevStandaloneChainKey(),
+		PendingPacketsIndexKey(),
 		BouncingSlashKey(),
 		RetryAllowedKey(),
 	}