+++ x/ccv/types/ccv.go
@@ -111,3 +111,17 @@ var ( // slice types can't be const
 	SlashPacketHandledResult = PacketAckResult([]byte{byte(2)})
 	SlashPacketBouncedResult = PacketAckResult([]byte{byte(3)})
 )
+
+// An exported wrapper around the auto generated isConsumerPacketData_Data interface, only for
+// AppendPendingPacket to accept the interface as an argument.
+type ExportedIsConsumerPacketData_Data interface {
+	isConsumerPacketData_Data
+}
+
+func NewConsumerPacketData(cpdType ConsumerPacketDataType, data isConsumerPacketData_Data, idx uint64) ConsumerPacketData {
+	return ConsumerPacketData{
+		Type: cpdType,
+		Data: data,
+		Idx:  idx,
+	}
+}